install.sql: perform the substitutions for autoconf
Another fix for TODO .

The provided fix from TODO was targetted as cmake, but in the case of
autoconf (configure && make), the variables were not substituted in the
install.sql, so when installed, install.sql contains:

    source @MRN_DATA_DIR@/install_plugin.sql
    source @MRN_DATA_DIR@/update.sql

( @MRN_DATA_DIR@ are not substituted ).

According to https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.71/html_node/Installation-Directory-Variables.html
we can not simply use AC_SUBST to let configure replace MRN_DATA_DIR and
the replacement needs to be done by make, so that a different directory
can be selected during make install.