data/install.sql: substitute @MRN_DATA_DIR@ for autoconf build
In the case of autoconf (configure && make), the variables were not
substituted in the install.sql, so after installation, install.sql
contains:

    source @MRN_DATA_DIR@/install_plugin.sql
    source @MRN_DATA_DIR@/update.sql

( @MRN_DATA_DIR@ are not substituted ).

According to https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.71/html_node/Installation-Directory-Variables.html
we can not simply use AC_SUBST to let configure substitute MRN_DATA_DIR
and the substitution needs to be done by make, so that a different
directory can be selected during make install.

To comply with this, define a rule to create install.sql from
install.sql.in and reorganise the target so that install.sql does not
get included in the distribution created by make dist